+++ lib/br_danfe/danfe.rb
@@ -3,7 +3,7 @@ module BrDanfe
     attr_reader :options
 
     def initialize(xml)
-      @xml = DanfeLib::XML.new(xml)
+      @xml = BrDanfe::XML.new(xml)
       @pdf = DanfeLib::Document.new
       @options = BrDanfe::Logo::Config.new
