+++ spec/lib/danfe_lib/icmstot_spec.rb
@@ -5,7 +5,7 @@ describe BrDanfe::DanfeLib::Icmstot do
   let(:output_pdf) { "#{base_dir}output.pdf" }
 
   let(:pdf) { BrDanfe::DanfeLib::Document.new }
-  let(:xml) { BrDanfe::DanfeLib::XML.new(xml_as_string) }
+  let(:xml) { BrDanfe::XML.new(xml_as_string) }
 
   subject { described_class.new(pdf, xml) }
